fix: Vuln/AssetExportStatus.status stores string
When deserializing from JSON to Vuln/AssetExportStatus dataclass, the
status field is actually a string and not automatically converted to
ExportRequestStatus. When checking the status, it should be compared to
the raw value of ExportRequestStatus via ExportRequestStatus.value
instead of the enum member itself, otherwise the comparison will always
fail.