fix: use varchar255 instead of varchar34 to store UUIDs
For some peculiar reason, values that are claimed to be UUIDs by Tenable
can actually be longer than the 36 characters which means they aren't
actually UUIDs. (I specified varchar34 originally because I messed up) I
decided to use VARCHAR(255) instead just to be on the safe side to
account for this situation (which should not have happend in the first
place). Thanks Tenable for not following standards!